# utils/result: optimize result_parallel_for_each