@avikivity 2022-02-09
Using a coroutine means that if everything returns immediately then we pay the cost of an allocation that we would not otherwise need.

I think for our use case it is okay. Even when using CL=ONE and the coordinator is co-located with the replica, they are on separate execution stages and so the result is never immediately ready.