@piodul 2022-02-10
v1 -> v1.1: (previous branch [link](https://github.com/piodul/scylla/tree/improve-result-parallel-for-each-v1))

- Moved `result_with_exception` from `utils::internal` to `utils`